% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_complete_deliveries}
\alias{metrc_complete_deliveries}
\title{Complete Deliveries}
\usage{
metrc_complete_deliveries(license_number, deliveries)
}
\description{
Complete Deliveries
}
\note{
See \url{https://api-or.metrc.com/Documentation/#Sales.put_sales_v1_deliveries_complete}
}
