% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_delete_delivery}
\alias{metrc_delete_delivery}
\title{Delete Delivery}
\usage{
metrc_delete_delivery(license_number, id)
}
\description{
Delete Delivery
}
\note{
See \url{https://api-or.metrc.com/Documentation/#Sales.delete_sales_v1_delivery_{id}}
}
