% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{metrc_delete_item}
\alias{metrc_delete_item}
\title{Delete Items}
\usage{
metrc_delete_item(license_number, id)
}
\arguments{
\item{license_number}{Facility license number}

\item{id}{Item ID}
}
\description{
Delete Items
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Items.delete_items_v1_{id}}
}
