% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{metrc_delete_patient}
\alias{metrc_delete_patient}
\title{Delete Patient}
\usage{
metrc_delete_patient(license_number, id)
}
\description{
Delete Patient
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Patients.delete_patients_v1_{id}}
}
