% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{metrc_delete_room}
\alias{metrc_delete_room}
\title{Delete room}
\usage{
metrc_delete_room(license_number, id)
}
\description{
Delete room
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Rooms.delete_rooms_v1_{id}}
}
