% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_delete_sales_receipt}
\alias{metrc_delete_sales_receipt}
\title{Delete Receipt}
\usage{
metrc_delete_sales_receipt(license_number, id)
}
\description{
Delete Receipt
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Sales.delete_sales_v1_receipts_{id}}
}
