% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strains.R
\name{metrc_delete_strain}
\alias{metrc_delete_strain}
\title{Delete Strain}
\usage{
metrc_delete_strain(license_number, id)
}
\description{
Delete Strain
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Strains.delete_strains_v1_{id}}
}
