% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_get_adjust_reasons}
\alias{metrc_get_adjust_reasons}
\title{Get Package Adjustment Reasons}
\usage{
metrc_get_adjust_reasons(licenseNumber = NULL)
}
\description{
Get Package Adjustment Reasons
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.get_packages_v1_adjust_reasons}
}
