% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_get_deliveries}
\alias{metrc_get_deliveries}
\alias{metrc_get_deliveries}
\title{Get Deliveries}
\usage{
metrc_get_deliveries(license_number, delivery_id)

metrc_get_deliveries(license_number, delivery_id)
}
\description{
Get Deliveries

Get Delivery Info
}
\note{
See \url{https://api-or.metrc.com/Documentation/#Sales.get_sales_v1_deliveries}

See \url{https://api-or.metrc.com/Documentation/#Sales.get_sales_v1_deliveries}
}
