% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_get_delivery}
\alias{metrc_get_delivery}
\title{Get Delivery Info}
\usage{
metrc_get_delivery(license_number, delivery_id)
}
\description{
Get Delivery Info
}
\note{
See \url{https://api-or.metrc.com/Documentation/#Sales.get_sales_v1_deliveries}
}
