% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facilites.R
\name{metrc_get_facilites}
\alias{metrc_get_facilites}
\title{Get Facilities Data}
\usage{
metrc_get_facilites()
}
\description{
Get Facilities Data
}
\note{
See url{https://api-co.metrc.com/Documentation/#Facilities.get_facilities_v1}
}
