% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facilites.R
\name{metrc_get_facilities}
\alias{metrc_get_facilities}
\title{Get Facilities Data}
\usage{
metrc_get_facilities()
}
\value{
tibble Containing information about each facility. License is a list-col with info about the license.
}
\description{
Gives information about all facilities user can access.
}
\note{
See url{https://api-co.metrc.com/Documentation/#Facilities.get_facilities_v1}
}
