% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvests.R
\name{metrc_get_harvest}
\alias{metrc_get_harvest}
\title{Get Individual Harvest Data}
\usage{
metrc_get_harvest(id)
}
\arguments{
\item{id}{Plant ID}
}
\description{
Get Individual Harvest Data
}
\note{
See url{https://api-co.metrc.com/Documentation/#Harvests.get_harvests_v1_{id}}
}
