% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvests.R
\name{metrc_get_harvests_active}
\alias{metrc_get_harvests_active}
\title{Get Active Harvests of Facility}
\usage{
metrc_get_harvests_active(license_number)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Get Active Harvests of Facility
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Harvests.get_harvests_v1_active}
}
