% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvests.R
\name{metrc_get_harvests_inactive}
\alias{metrc_get_harvests_inactive}
\title{Get Inactive Harvests of Facility}
\usage{
metrc_get_harvests_inactive(license_number)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Get Inactive Harvests of Facility
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Harvests.get_harvests_v1_inactive}
}
