% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{metrc_get_item}
\alias{metrc_get_item}
\title{Get Item Data}
\usage{
metrc_get_item(id)
}
\arguments{
\item{id}{Item ID}
}
\description{
Get Item Data
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Items.get_items_v1_{id}}
}
