% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{metrc_get_items_active}
\alias{metrc_get_items_active}
\title{Get Active Items}
\usage{
metrc_get_items_active(license_number)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Get Active Items
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Items.get_items_v1_active}
}
