% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{metrc_get_items_categories}
\alias{metrc_get_items_categories}
\title{Get Items Categories}
\usage{
metrc_get_items_categories()
}
\description{
Get Items Categories
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Items.get_items_v1_categories}
}
