% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_tests.R
\name{metrc_get_lab_test_states}
\alias{metrc_get_lab_test_states}
\title{Get Lab Test States}
\usage{
metrc_get_lab_test_states()
}
\description{
Get Lab Test States
}
\note{
See \url{https://api-co.metrc.com/Documentation/#LabTests.get_labtests_v1_states}
}
