% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_tests.R
\name{metrc_get_lab_test_types}
\alias{metrc_get_lab_test_types}
\title{Get Lab Test Types}
\usage{
metrc_get_lab_test_types()
}
\description{
Get Lab Test Types
}
\note{
See \url{https://api-co.metrc.com/Documentation/#LabTests.get_labtests_v1_types}
}
