% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_get_package}
\alias{metrc_get_package}
\title{Get Package Info}
\usage{
metrc_get_package(id)
}
\arguments{
\item{id}{ID of package}
}
\description{
Get Package Info
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.get_packages_v1_{id}}
}
