% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_get_packages_active}
\alias{metrc_get_packages_active}
\title{Get Active Packages}
\usage{
metrc_get_packages_active(license_number)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Get Active Packages
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.get_packages_v1_active}
}
