% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_get_packages_inactive}
\alias{metrc_get_packages_inactive}
\title{Get Inactive Packages}
\usage{
metrc_get_packages_inactive(license_number)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Get Inactive Packages
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.get_packages_v1_inactive}
}
