% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_get_packages_onhold}
\alias{metrc_get_packages_onhold}
\title{Get Onhold Packages}
\usage{
metrc_get_packages_onhold(license_number)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Get Onhold Packages
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.get_packages_v1_onhold}
}
