% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_get_packages_types}
\alias{metrc_get_packages_types}
\title{Get Package Types}
\usage{
metrc_get_packages_types()
}
\description{
Get Package Types
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.get_packages_v1_types}
}
