% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{metrc_get_patient}
\alias{metrc_get_patient}
\title{Get Patient Info}
\usage{
metrc_get_patient(id)
}
\arguments{
\item{id}{ID of patient}
}
\description{
Get Patient Info
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Patients.get_patients_v1_{id}}
}
