% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{metrc_get_patient_status}
\alias{metrc_get_patient_status}
\title{Get Patient Status}
\usage{
metrc_get_patient_status(license_number, patient_license_number)
}
\description{
Get Patient Status
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Patients.get_patients_v1_status_{patientLicenseNumber}}
}
