% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{metrc_get_patients_active}
\alias{metrc_get_patients_active}
\title{Get Active Patients}
\usage{
metrc_get_patients_active(license_number)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Get Active Patients
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Patients.get_patients_v1_active}
}
