% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{metrc_get_plant}
\alias{metrc_get_plant}
\title{Get Plant Info}
\usage{
metrc_get_plant(id)
}
\description{
Get Plant Info
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Plants.get_plants_v1_{id}}
}
