% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_batches.R
\name{metrc_get_plant_batch}
\alias{metrc_get_plant_batch}
\title{Get Plant Batches}
\usage{
metrc_get_plant_batch(id)
}
\description{
Get Plant Batches
}
\note{
See \url{https://api-co.metrc.com/Documentation/#PlantBatches.get_plantbatches_v1_{id}}
}
