% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_batches.R
\name{metrc_get_plant_batchs_active}
\alias{metrc_get_plant_batchs_active}
\title{Get Active Plant Batches}
\usage{
metrc_get_plant_batchs_active(license_number)
}
\description{
Get Active Plant Batches
}
\note{
See \url{https://api-co.metrc.com/Documentation/#PlantBatches.get_plantbatches_v1_active}
}
