% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_batches.R
\name{metrc_get_plant_batchs_inactive}
\alias{metrc_get_plant_batchs_inactive}
\title{Get Inactive Plant Batches}
\usage{
metrc_get_plant_batchs_inactive(license_number)
}
\description{
Get Inactive Plant Batches
}
\note{
See \url{https://api-co.metrc.com/Documentation/#PlantBatches.get_plantbatches_v1_inactive}
}
