% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_batches.R
\name{metrc_get_plant_batchs_types}
\alias{metrc_get_plant_batchs_types}
\title{Get Plant Batch Types}
\usage{
metrc_get_plant_batchs_types()
}
\description{
Get Plant Batch Types
}
\note{
See \url{https://api-co.metrc.com/Documentation/#PlantBatches.get_plantbatches_v1_types}
}
