% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{metrc_get_plants_flowering}
\alias{metrc_get_plants_flowering}
\title{Get Flowering Plants}
\usage{
metrc_get_plants_flowering(license_number)
}
\description{
Get Flowering Plants
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Plants.get_plants_v1_vegetative}
}
