% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{metrc_get_plants_inactive}
\alias{metrc_get_plants_inactive}
\title{Get Inactive Plants}
\usage{
metrc_get_plants_inactive(license_number)
}
\description{
Get Inactive Plants
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Plants.get_plants_v1_inactive}
}
