% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{metrc_get_plants_onhold}
\alias{metrc_get_plants_onhold}
\title{Get Onhold Plants}
\usage{
metrc_get_plants_onhold(license_number)
}
\description{
Get Onhold Plants
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Plants.get_plants_v1_onhold}
}
