% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{metrc_get_plants_vegetative}
\alias{metrc_get_plants_vegetative}
\title{Get Vegetative Plants}
\usage{
metrc_get_plants_vegetative(license_number)
}
\description{
Get Vegetative Plants
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Plants.get_plants_v1_vegetative}
}
