% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_get_return_reasons}
\alias{metrc_get_return_reasons}
\title{Return Reasons}
\usage{
metrc_get_return_reasons()
}
\description{
Return Reasons
}
\note{
See \url{https://api-or.metrc.com/Documentation/#Sales.get_sales_v1_delivery_returnreasons}
}
