% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{metrc_get_rooms_active}
\alias{metrc_get_rooms_active}
\title{Get Active Rooms}
\usage{
metrc_get_rooms_active(license_number)
}
\description{
Get Active Rooms
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Rooms.get_rooms_v1_active}
}
