% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_get_sales_customer_types}
\alias{metrc_get_sales_customer_types}
\title{Get Customer Types}
\usage{
metrc_get_sales_customer_types()
}
\description{
Get Customer Types
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Sales.get_sales_v1_customertypes}
}
