% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_get_sales_receipt}
\alias{metrc_get_sales_receipt}
\title{Get Receipt}
\usage{
metrc_get_sales_receipt(id)
}
\description{
Get Receipt
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Sales.get_sales_v1_receipts_{id}}
}
