% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_get_sales_receipts}
\alias{metrc_get_sales_receipts}
\title{Get Sales Receipts}
\usage{
metrc_get_sales_receipts(license_number)
}
\description{
Get Sales Receipts
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Sales.get_sales_v1_receipts}
}
