% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_get_sales_transactions}
\alias{metrc_get_sales_transactions}
\title{Get Sales Transactions}
\usage{
metrc_get_sales_transactions(license_number)
}
\description{
Get Sales Transactions
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Sales.get_sales_v1_transactions}
}
