% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_get_sales_transactions_date}
\alias{metrc_get_sales_transactions_date}
\title{Get Sales Transactions by Date}
\usage{
metrc_get_sales_transactions_date(license_number, date)
}
\description{
Get Sales Transactions by Date
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Sales.get_sales_v1_transactions_{date}}
}
