% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strains.R
\name{metrc_get_strain}
\alias{metrc_get_strain}
\title{Get Strain}
\usage{
metrc_get_strain(id)
}
\description{
Get Strain
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Strains.get_strains_v1_{id}}
}
