% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strains.R
\name{metrc_get_strains_active}
\alias{metrc_get_strains_active}
\title{Get Active Strains}
\usage{
metrc_get_strains_active(license_number)
}
\description{
Get Active Strains
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Strains.get_strains_v1_active}
}
