% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{metrc_get_transfers_delivery}
\alias{metrc_get_transfers_delivery}
\title{Get Deliveries}
\usage{
metrc_get_transfers_delivery(id)
}
\description{
Get Deliveries
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Transfers.get_transfers_v1_{id}_deliveries}
}
