% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{metrc_get_transfers_incoming}
\alias{metrc_get_transfers_incoming}
\title{Get Incoming Transfers}
\usage{
metrc_get_transfers_incoming(license_number)
}
\description{
Get Incoming Transfers
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Transfers.get_transfers_v1_incoming}
}
