% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{metrc_get_transfers_package_states}
\alias{metrc_get_transfers_package_states}
\title{Get Package States}
\usage{
metrc_get_transfers_package_states()
}
\description{
Get Package States
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Transfers.get_transfers_v1_delivery_{id}_packages}
}
