% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{metrc_get_transfers_packages}
\alias{metrc_get_transfers_packages}
\alias{metrc_get_transfers_packages}
\title{Get Packages}
\usage{
metrc_get_transfers_packages()

metrc_get_transfers_packages()
}
\description{
Get Packages

Get Package States
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Transfers.get_transfers_v1_{id}_deliveries}

See \url{https://api-co.metrc.com/Documentation/#Transfers.get_transfers_v1_delivery_{id}_packages}
}
