% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{metrc_get_transfers_rejected}
\alias{metrc_get_transfers_rejected}
\title{Get Rejected Transfers}
\usage{
metrc_get_transfers_rejected(license_number)
}
\description{
Get Rejected Transfers
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Transfers.get_transfers_v1_rejected}
}
