% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_post_deliveries}
\alias{metrc_post_deliveries}
\title{Post Deliveries}
\usage{
metrc_post_deliveries(license_number, deliveries)
}
\description{
Post Deliveries
}
\note{
See \url{https://api-or.metrc.com/Documentation/#Sales.post_sales_v1_deliveries}
}
