% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvests.R
\name{metrc_post_harvests_createpackage}
\alias{metrc_post_harvests_createpackage}
\title{Post New Harvest}
\usage{
metrc_post_harvests_createpackage(license_number, harvest, item, weight,
  unit_of_weight, tag, is_production_batch, production_batch_number,
  product_requires_remediation, remediate_product, remediation_method_id,
  remediation_date, remediation_steps, actual_date)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Post New Harvest
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Harvests.post_harvests_v1_createpackages}
}
