% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvests.R
\name{metrc_post_harvests_finish}
\alias{metrc_post_harvests_finish}
\title{Post Finish Harvest}
\usage{
metrc_post_harvests_finish(license_number, id, actual_date)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Post Finish Harvest
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Harvests.post_harvests_v1_finish}
}
