% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvests.R
\name{metrc_post_harvests_removewaste}
\alias{metrc_post_harvests_removewaste}
\title{Post Harvest Waste}
\usage{
metrc_post_harvests_removewaste(license_number, id, unit_of_weight,
  waste_weight, actual_date)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Post Harvest Waste
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Harvests.post_harvests_v1_removewaste}
}
