% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvests.R
\name{metrc_post_harvests_unfinish}
\alias{metrc_post_harvests_unfinish}
\title{Post Unfinish Harvest}
\usage{
metrc_post_harvests_unfinish(license_number, id)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Post Unfinish Harvest
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Harvests.post_harvests_v1_unfinish}
}
