% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{metrc_post_items_create}
\alias{metrc_post_items_create}
\title{Post New Items}
\usage{
metrc_post_items_create(license_number, item_category, name,
  unit_of_measurement, strain, unit_thc_content,
  unit_thc_content_unit_of_measure, unit_weight, unit_weight_unit_of_measure)
}
\arguments{
\item{license_number}{Facility license number}
}
\description{
Post New Items
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Items.post_items_v1_create}
}
