% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_tests.R
\name{metrc_post_lab_test}
\alias{metrc_post_lab_test}
\title{Post Lab Tests}
\usage{
metrc_post_lab_test(license_number, results)
}
\description{
Post Lab Tests
}
\note{
See \url{https://api-co.metrc.com/Documentation/#LabTests.post_labtests_v1_record}
}
