% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_post_packages_adjust}
\alias{metrc_post_packages_adjust}
\title{Post Adjust Package}
\usage{
metrc_post_packages_adjust(license_number, label, quantity, unit_of_measure,
  adjustment_reason, adjustment_date, reason_note)
}
\description{
Post Adjust Package
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.post_packages_v1_adjust}
}
