% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_post_packages_change}
\alias{metrc_post_packages_change}
\title{Post Change Package}
\usage{
metrc_post_packages_change(license_number, label, item)
}
\description{
Post Change Package
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.post_packages_v1_change_item}
}
