% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_post_packages_create}
\alias{metrc_post_packages_create}
\title{Post New Package}
\usage{
metrc_post_packages_create(license_number, tag, item, quantity, unit_of_measure,
  is_production_batch, production_batch_number, product_requires_remediation,
  actual_date, ingredients)
}
\arguments{
\item{ingredients}{list-col}
}
\description{
Post New Package
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.post_packages_v1_create}
}
