% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_post_packages_finish}
\alias{metrc_post_packages_finish}
\title{Post Finish Package}
\usage{
metrc_post_packages_finish(license_number, label, actual_date)
}
\description{
Post Finish Package
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.post_packages_v1_finish}
}
