% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_post_packages_remediate}
\alias{metrc_post_packages_remediate}
\title{Post Remediate Package}
\usage{
metrc_post_packages_remediate(
  license_number,
  package_label,
  remediation_method_name,
  remediation_date,
  remediation_steps
)
}
\description{
Post Remediate Package
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.post_packages_v1_remediate}
}
