% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{metrc_post_packages_unfinish}
\alias{metrc_post_packages_unfinish}
\title{Post Unfinish Package}
\usage{
metrc_post_packages_unfinish(license_number, label)
}
\description{
Post Unfinish Package
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Packages.post_packages_v1_unfinish}
}
