% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{metrc_post_patients_update}
\alias{metrc_post_patients_update}
\title{Post Update Patient}
\usage{
metrc_post_patients_update(
  license_number,
  patient_license_number,
  new_license_number,
  license_effective_start_date,
  license_effective_end_date,
  recommended_plants,
  recommended_smokable_quantity,
  actual_date
)
}
\description{
Post Update Patient
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Patients.post_patients_v1_update}
}
