% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_batches.R
\name{metrc_post_plant_batches_change_growth_phase}
\alias{metrc_post_plant_batches_change_growth_phase}
\title{Post Change Plant Growth Phase}
\usage{
metrc_post_plant_batches_change_growth_phase(license_number, id, count,
  starting_tag, growth_phase, new_room, growth_date)
}
\description{
Post Change Plant Growth Phase
}
\note{
See \url{https://api-co.metrc.com/Documentation/#PlantBatches.post_plantbatches_v1_changegrowthphase}
}
