% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_batches.R
\name{metrc_post_plant_batches_create_packages}
\alias{metrc_post_plant_batches_create_packages}
\title{Post Create Packages}
\usage{
metrc_post_plant_batches_create_packages(
  license_number,
  id,
  item,
  tag,
  count,
  actual_date
)
}
\description{
Post Create Packages
}
\note{
See \url{https://api-co.metrc.com/Documentation/#PlantBatches.post_plantbatches_v1_createpackages}
}
