% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_batches.R
\name{metrc_post_plant_batches_destroy}
\alias{metrc_post_plant_batches_destroy}
\title{Post Destroy Plant}
\usage{
metrc_post_plant_batches_destroy(
  license_number,
  id,
  count,
  reason_note,
  actual_date
)
}
\description{
Post Destroy Plant
}
\note{
See \url{https://api-co.metrc.com/Documentation/#PlantBatches.post_plantbatches_v1_destroy}
}
