% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_batches.R
\name{metrc_post_plant_batchs_create_plantings}
\alias{metrc_post_plant_batchs_create_plantings}
\title{Post Create Plantings}
\usage{
metrc_post_plant_batchs_create_plantings(license_number, name, type, count,
  strain, actual_date)
}
\description{
Post Create Plantings
}
\note{
See \url{https://api-co.metrc.com/Documentation/#PlantBatches.post_plantbatches_v1_createplantings}
}
