% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{metrc_post_plants_change_growth_phase}
\alias{metrc_post_plants_change_growth_phase}
\title{Post Move Plants}
\usage{
metrc_post_plants_change_growth_phase(
  license_number,
  id,
  label,
  new_tag,
  growth_phase,
  new_room,
  growth_date
)
}
\description{
Post Move Plants
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Plants.post_plants_v1_moveplants}
}
