% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{metrc_post_plants_create_plantings}
\alias{metrc_post_plants_create_plantings}
\title{Post New Plant Plantings}
\usage{
metrc_post_plants_create_plantings(
  license_number,
  plant_label,
  plant_batch_name,
  plant_batch_type,
  plant_count,
  strain_name,
  actual_date
)
}
\description{
Post New Plant Plantings
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Plants.post_plants_v1_create_plantings}
}
