% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{metrc_post_plants_destroy}
\alias{metrc_post_plants_destroy}
\title{Post Delete Plants}
\usage{
metrc_post_plants_destroy(license_number, id, label, reason_note, actual_date)
}
\description{
Post Delete Plants
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Plants.post_plants_v1_destroyplants}
}
