% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{metrc_post_plants_harvest}
\alias{metrc_post_plants_harvest}
\title{Post Harvest Plants}
\usage{
metrc_post_plants_harvest(
  license_number,
  plant,
  weight,
  unit_of_weight,
  drying_room,
  harvest_name,
  actual_date
)
}
\description{
Post Harvest Plants
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Plants.post_plants_v1_harvestplants}
}
