% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{metrc_post_plants_manicure}
\alias{metrc_post_plants_manicure}
\title{Post Manicure Plants}
\usage{
metrc_post_plants_manicure(license_number, plant, weight, unit_of_weight,
  drying_room, harvest_name, actual_date)
}
\description{
Post Manicure Plants
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Plants.post_plants_v1_manicureplants}
}
