% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{metrc_post_plants_move_plants}
\alias{metrc_post_plants_move_plants}
\title{Post Move Plants}
\usage{
metrc_post_plants_move_plants(license_number, id, label, room, actual_date)
}
\description{
Post Move Plants
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Plants.post_plants_v1_moveplants}
}
