% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{metrc_post_rooms}
\alias{metrc_post_rooms}
\title{Post New Room}
\usage{
metrc_post_rooms(license_number, name)
}
\description{
Post New Room
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Rooms.post_rooms_v1_create}
}
