% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{metrc_post_rooms_update}
\alias{metrc_post_rooms_update}
\title{Post Update Room}
\usage{
metrc_post_rooms_update(license_number, id, name)
}
\description{
Post Update Room
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Rooms.post_rooms_v1_update}
}
