% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_post_sales_receipt}
\alias{metrc_post_sales_receipt}
\title{Post Receipts}
\usage{
metrc_post_sales_receipt(license_number, sales)
}
\description{
Post Receipts
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Sales.post_sales_v1_receipts}
}
