% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strains.R
\name{metrc_post_strains}
\alias{metrc_post_strains}
\title{Post Strains}
\usage{
metrc_post_strains(
  license_number,
  name,
  testing_status,
  thc_level,
  cbd_level,
  indica_percentage,
  sativa_percentage
)
}
\description{
Post Strains
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Strains.post_strains_v1_create}
}
