% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strains.R
\name{metrc_post_strains_update}
\alias{metrc_post_strains_update}
\title{Post Update Strains}
\usage{
metrc_post_strains_update(
  license_number,
  id,
  name,
  testing_status,
  thc_level,
  cbd_level,
  indica_percentage,
  sativa_percentage
)
}
\description{
Post Update Strains
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Strains.post_strains_v1_update}
}
