% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_put_deliveries}
\alias{metrc_put_deliveries}
\title{Put Deliveries}
\usage{
metrc_put_deliveries(license_number, deliveries)
}
\description{
Put Deliveries
}
\note{
See \url{https://api-or.metrc.com/Documentation/#Sales.put_sales_v1_deliveries}
}
