% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_put_sales_receipt}
\alias{metrc_put_sales_receipt}
\title{Put Receipts}
\usage{
metrc_put_sales_receipt(license_number, sales)
}
\description{
Put Receipts
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Sales.put_sales_v1_receipts}
}
