% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{metrc_put_sales_transactions_date}
\alias{metrc_put_sales_transactions_date}
\title{Put Sales Transactions by Date}
\usage{
metrc_put_sales_transactions_date(
  license_number,
  date,
  package_label,
  quantity,
  unit_of_measure,
  total_amount
)
}
\description{
Put Sales Transactions by Date
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Sales.post_sales_v1_transactions_{date}}
}
