% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employees.R
\name{metrc_get_employees}
\alias{metrc_get_employees}
\title{Get Employee Data}
\usage{
metrc_get_employees(license_number)
}
\arguments{
\item{license_number}{Employee license number}
}
\value{
data.frame Containing information about the employess of a facility
}
\description{
This provides information about all employees at a specified facility.
}
\note{
See \url{https://api-co.metrc.com/Documentation/#Employees.get_employees_v1}
}
