% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{metrc_get_transfers_packages}
\alias{metrc_get_transfers_packages}
\title{Get Packages}
\usage{
metrc_get_transfers_packages(id)
}
\description{
Get Packages
}
\note{
See \url{https://api-or.metrc.com/Documentation/#Transfers.get_transfers_v1_delivery_{id}_packages}
}
